got groups to load on initial render by having the client wait until they are fetched from the api
